add argument for ProductFragment, change code due to comments

+++ app/src/main/java/com/jsp/freshcartshop/viewmodel/MainViewModel.kt
@@ -37,7 +37,7 @@ class MainViewModel : ViewModel() {
         }
     }
 
-    fun loadProduct(id : Int) {
+    fun loadProduct(id : Long) {
         viewModelScope.launch {
             try {
                 val response = freshCartRepository.getProductById(id)
